feat!: Genesis Accounts can not contain empty code
Genesis accounts currently serve no purpose for non-contract accounts.
Currently, they can exist but code and storage will be empty and
therefore has no effect besids ensuring the matching account is subject
to similar validations as a contract.

This change restrics valid genesis accounts to those with Code only.
Storage can still be empty for contracts, as there is no requirement for
values to be set in state for all contracts.