fix!: Add back code hash check for empty code
On some chains running ethermint, a self destruct bug that caused the
code for other contract accounts to be deleted would result in exported
genesis accounts with empty code.  A patch was then added to genesis
that allowed the hash check to be skipped in order for these genesis
files to be re-imported successfully.

This does not affect us contracts on Kava, so this check is being
removed.