refactor: Remove precompile genesis checks
This completely removes the dependency on the go-ethereum precompile
registry in genesis and relaxes the precompile code check on init
genesis to remove uneeded coupling between modules and interfaces.

The x/evm module should be largely unaware of the inner workings of the
EVM and Kava extensions to it for stateful precompiles.  This is a
result of a couple design decisions, such as not automatically including
precompiles in the access list and creating an API that allows Kava to
have full control over the code set for precompile state and future
versioning.

One of these design decisions is that etheremint should not validate
enabled precompiles at all (still validates code is set), but
leave it to the runtime execution of them to detect invalid states and
panic.

This is still the open question if the x/evm module should have enabled
precompiles at all (and it probably shouldn't), but this is out of scope
for now and will be revisited.